Improve Dymos convergence
Switch from Gauss-Lobatto transcription to Radau Pseudospectral
transcription, which is preferred when using a differential inclusion
approach.

Enable solve_segments for the objective state variable and fix the
initial state for all variables.. This change causes a nonlinear solver
to be used and if nonconvergence is detected pyoptsparse will backtrack
along its line search.

The motivation for these changes came from
https://openmdao.github.io/dymos/examples/commercial_aircraft.html